Clean up a few warnings.

+++ neo/framework/File_SaveGame.h
@@ -132,7 +132,7 @@ public:
 	virtual int				Write( const void * buffer, int len );
 
 	// this file is strictly streaming, you can't seek at all
-	virtual int				Length() const  { return compressedLength; }
+	virtual int				Length() const  { return (int)compressedLength; }
 	virtual void			SetLength( size_t len ) { compressedLength = len; }
 	virtual int				Tell() const { assert( 0 ); return 0; }
 	virtual int				Seek( long offset, fsOrigin_t origin ) { assert( 0 ); return 0; }
